[moe refactor] fix unit import error
