+++ colossalai/moe/load_balance.py
@@ -7,8 +7,8 @@ from torch import Tensor, nn
 from torch.distributed import ProcessGroup
 
 from colossalai.cluster import ProcessGroupMesh
-from colossalai.moe.experts import MLPExperts
 from colossalai.moe.manager import MOE_MANAGER
+from colossalai.shardformer.layer.moe.layers import MLPExperts
 from colossalai.zero.low_level import LowLevelZeroOptimizer
