+++ tests/test_moe/test_grad_handler.py
@@ -5,8 +5,8 @@ import torch.nn as nn
 
 import colossalai
 from colossalai.accelerator import get_accelerator
-from colossalai.moe import SparseMLP
 from colossalai.moe.manager import MOE_MANAGER
+from colossalai.shardformer.layer.moe.layers import SparseMLP
 from colossalai.testing import assert_equal_in_group, rerun_if_address_is_in_use, spawn
 from tests.test_moe.moe_utils import MoeGradientHandler
